Rename the jpa module to repository
Jpa is too implementation-specific to be in the module name
