+++ modules/repository/src/main/kotlin/ca/rmen/poetassistant/repository/definitions/DefinitionEntity.kt
@@ -17,28 +17,27 @@
  * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ca.rmen.poetassistant.jpa.rhymes
+package ca.rmen.poetassistant.repository.definitions
 
 import javax.persistence.Column
 import javax.persistence.EmbeddedId
 import javax.persistence.Entity
 import javax.persistence.Table
 
+/**
+ * Our embedded database doesn't have an id column.
+ * A word may have multiple definition entries for a given part_of_speech value.
+ * So, we have to use all three attributes as the primary key
+ */
 @Entity
-@Table(name = "word_variants")
-data class RhymerEntity(
+@Table(name = "dictionary")
+data class DefinitionEntity(
     @EmbeddedId
-    private val id: RhymerEntityPK = RhymerEntityPK(),
+    private val id: DefinitionEntityPK = DefinitionEntityPK(),
     @Column(insertable = false, updatable = false)
     val word: String = "",
     @Column(insertable = false, updatable = false)
-    val variantNumber: Int = 0,
+    val partOfSpeech: String = "",
     @Column(insertable = false, updatable = false)
-    val stressSyllables: String = "",
-    @Column(insertable = false, updatable = false)
-    val lastSyllable: String = "",
-    @Column(insertable = false, updatable = false)
-    val lastTwoSyllables: String = "",
-    @Column(insertable = false, updatable = false)
-    val lastThreeSyllables: String = "",
+    val definition: String = ""
 )
