+++ modules/repository/src/main/kotlin/ca/rmen/poetassistant/repository/thesaurus/ThesaurusEntityPK.kt
@@ -17,15 +17,15 @@
  * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ca.rmen.poetassistant.jpa.rhymes
+package ca.rmen.poetassistant.repository.thesaurus
 
-import org.springframework.data.repository.CrudRepository
+import java.io.Serializable
+import javax.persistence.Embeddable
 
-interface RhymerRepository : CrudRepository<RhymerEntity, RhymerEntityPK> {
-    fun findAllByWord(word: String): List<RhymerEntity>
-
-    // These long method names are possible thanks to jpa:
-    // https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#jpa.query-methods.query-creation
-    fun findByStressSyllablesAndWordNotOrderByWord(stressSyllables: String, word: String): List<RhymerEntity>
-
-}
+@Embeddable
+data class ThesaurusEntityPK(
+    val word: String = "",
+    val wordType: String = "",
+    val synonyms: String = "",
+    val antonyms: String = ""
+) : Serializable
