+++ modules/repository/src/main/kotlin/ca/rmen/poetassistant/repository/wotd/StemEntity.kt
@@ -17,10 +17,23 @@
  * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ca.rmen.poetassistant.jpa.thesaurus
+package ca.rmen.poetassistant.repository.wotd
 
-import org.springframework.data.repository.CrudRepository
+import javax.persistence.Column
+import javax.persistence.Entity
+import javax.persistence.Id
+import javax.persistence.Table
 
-interface ThesaurusRepository : CrudRepository<ThesaurusEntity, ThesaurusEntityPK> {
-    fun findAllByWord(word: String): List<ThesaurusEntity>
+@Entity
+@Table(name = "stems")
+class StemEntity {
+    @Id
+    @Column(insertable = false, updatable = false)
+    val word: String = ""
+
+    @Column(insertable = false, updatable = false)
+    val stem: String = ""
+
+    @Column(name = "google_ngram_frequency", insertable = false, updatable = false)
+    val googleNgramFrequency: Int = 0
 }
