+++ modules/service/build.gradle.kts
@@ -6,5 +6,5 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-json")
 
     implementation(project(":modules:model"))
-    implementation(project(":modules:jpa"))
+    implementation(project(":modules:repository"))
 }
